Moved Commands to commands
