+++ test/test_webdriver_utils.py
@@ -1,5 +1,5 @@
 from openwpm import TaskManager
-from openwpm.Commands.utils.webdriver_utils import parse_neterror
+from openwpm.commands.utils.webdriver_utils import parse_neterror
 from openwpm.utilities import db_utils
 
 from .openwpmtest import OpenWPMTest
